ai_part/ocular_diagnosis_tabular_ai_system/requirements.txt
# FILE: requirements.txt
tensorflow>=2.14.0
pandas>=2.1.0
scikit-learn>=1.4.0
pyyaml>=6.0.1
matplotlib>=3.8.0
seaborn>=0.13.1
openpyxl>=3.1.2
tqdm>=4.67.0
pytest>=7.5.0
pyarrow>=13.0.0  
pydantic>=2.0.0